EAMxx: move DefaultMetadata dictionary to share/util
